Azure Serverless Function/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pandas
tweepy
keras==2.6
keras_preprocessing 
tensorflow==2.8.0
numpy
scikit-learn
pyodbc